# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Para o administrador do site  de Pet Shop, que precisa manter um controle e conhecimento do estoque  , o Shopping dos Bichos é um sistema de gerenciamento e controle de estoque, que permite cadastrar e editar todos os itens e serviços da loja.",
    "version": "1.0.0",
    "title": "Shopping dos Bichos",
    "contact": {
      "email": "shoppingdosbichos@gmail.com",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/product": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Finds products by status",
        "description": "Getting all products",
        "operationId": "findproducts",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/product"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "product"
        ],
        "summary": "Add a new product to the store",
        "description": "",
        "operationId": "addproduct",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "product object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/productCreate"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/product/{productId}": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Find product by ID",
        "description": "Returns a single product",
        "operationId": "getproductById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/product"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "product not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "product"
        ],
        "summary": "Deletes a product",
        "description": "",
        "operationId": "deleteproduct",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "product id to delete",
            "required": true,
            "type": "string",
            "format": "ObjectId"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "product not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "product"
        ],
        "summary": "Update an existing product",
        "description": "",
        "operationId": "updateproduct",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "string",
            "format": "ObjectId"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "product not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/pet": {
      "get": {
        "tags": [
          "pet"
        ],
        "summary": "Finds Pets by status",
        "description": "Getting all pets",
        "operationId": "findPets",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Pet"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "pet"
        ],
        "summary": "Add a new pet to the store",
        "description": "",
        "operationId": "addPet",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Pet object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Pet"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/pet/findByStatus/{status}": {
      "get": {
        "tags": [
          "pet"
        ],
        "summary": "Finds Pets by status",
        "description": "Getting pets for their status",
        "operationId": "findPetsByStatus",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "description": "Status values that need to be considered for filter",
            "required": true,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "available",
                "pending",
                "sold"
              ],
              "default": "available"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Pet"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/pet/{petId}": {
      "get": {
        "tags": [
          "pet"
        ],
        "summary": "Find pet by ID",
        "description": "Returns a single pet",
        "operationId": "getPetById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "petId",
            "in": "path",
            "description": "ID of pet to return",
            "required": true,
            "type": "string",
            "format": "ObjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Pet"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "pet"
        ],
        "summary": "Deletes a pet",
        "description": "",
        "operationId": "deletePet",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "petId",
            "in": "path",
            "description": "Pet id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "pet"
        ],
        "summary": "Update an existing pet",
        "description": "",
        "operationId": "updatePet",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "petId",
            "in": "path",
            "description": "ID of pet to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    },
    "BasicAuth": {
      "type": "basic"
    }
  },
  "definitions": {
    "id": {
      "type": "string",
      "format": "ObjectId",
      "example": "507f191e810c19729de860ea"
    },
...
    "DescriptionPet": {
      "type": "object",
      "properties": {
        "color": {
          "type": "string"
        },
        "weight": {
          "type": "number"
        },
        "size": {
          "type": "string"
        },
        "age": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Tag"
      }
    },
    "Pet": {
      "type": "object",
      "required": [
        "name",
        "photo_urls"
      ],
      "properties": {
        "id": {
          "$ref": "#/definitions/id"
        },
        "name": {
          "type": "string",
          "example": "doggie"
        },
        "type_pet": {
          "type": "string",
          "example": "dog"
        },
        "breed": {
          "type": "string",
          "example": "Pitbull"
        },
        "description": {
          "$ref": "#/definitions/DescriptionPet"
        },
        "photo_urls": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "in_adoption": {
          "type": "boolean",
          "example": false
        },
        "price": {
          "type": "number",
          "example": 1600
        },
        "status": {
          "type": "string",
          "description": "pet status in the store",
          "enum": [
            "available",
            "pending",
            "sold"
          ]
        }
      }
    },
    "productCreate": {
      "type": "object",
      "required": [
        "name",
        "photoUrl"
      ],
      "properties": {
        "name": {
          "type": "string",
          "example": "dorazel plus"
        },
        "description": {
          "type": "string",
          "example": "anti-garrapatas"
        },
        "category": {
          "type": "string",
          "example": "farmacia"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "photoUrl": {
          "type": "array",
          "items": {
            "type": "string",
            "example": "http://exaple.com/imageexample.jpg"
          }
        },
        "stock": {
          "type": "number",
          "example": 100
        },
        "expirationDate": {
          "type": "string",
          "format": "date"
        },
        "price": {
          "type": "number",
          "example": 1600
        },
        "status": {
          "type": "string",
          "description": "product status in the store",
          "enum": [
            "available",
            "pending",
            "sold"
          ]
        }
      }
    },
    "product": {
      "type": "object",
      "required": [
        "name",
        "photoUrl"
      ],
      "properties": {
        "id": {
          "$ref": "#/definitions/id"
        },
        "name": {
          "type": "string",
          "example": "dorazel plus"
        },
        "description": {
          "type": "string",
          "example": "anti-garrapatas"
        },
        "category": {
          "type": "string",
          "example": "farmacia"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "photoUrl": {
          "type": "array",
          "items": {
            "type": "string",
            "example": "http://exaple.com/imageexample.jpg"
          }
        },
        "stock": {
          "type": "number",
          "example": 100
        },
        "expirationDate": {
          "type": "string",
          "format": "date"
        },
        "price": {
          "type": "number",
          "example": 1600
        },
        "status": {
          "type": "string",
          "description": "product status in the store",
          "enum": [
            "available",
            "pending",
            "sold"
          ]
        }
      }
    },
...
  },
...
}